Move feature importances to BaseTimeSeriesForest
